iio: ad_adc: Refactor initialization
There are a couple of things with the current bindings of the ad_adc
driver that aren't quite optimal.  The 'adc-name-id' works as a second
level compatible attribute, a better solution is to directly put the name
in the compatible attribute. Also the name for the DMA stream is
"adc-adc-dma" while it should be just "rx" to be spec compliant. Also there
shouldn't be a property to indicate whether we want streaming or non
streaming DMA, this is was just a workaround for a broken DMA controller.

This patch factors the old probe functionality with the deprecated
device-tree bindings into a separate function in order to keep backwards
compatibility for now. Later patches will add support for new devices which
will use proper bindings.

While we are at it also switch to a table based device info.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>