iio: imu: adis16480: Validate the CRC when burst read function is used
The CRC code is part of the batch of data made available via the burst read
function. After the BURST_ID, the next 15 sequences of 16-bits represent
the data and the last two are the lower and upper CRC codes.
The crc code is validated by using the crc32 kernel library and the
result is pushed to buffers when the crc channel of IIO_FLAGS type is
enabled.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>
(cherry picked from commit 607f28b59b099f678e053b4769aae6b3c28cd599)