net: phy: adin: adjust compat layer to 4.19
This PHY layer was adjusted to build on 4.19, but wasn't tested until now.
The driver does seem to work, but phylib complains with a
warning-stack-trace that some calls to __mdiobus_{read,write} aren't locked
(and should be).

This happens in the phy_modify_mmd_changed() function, because the compat
layer re-implements it. When the kernel will be upgraded to 5.3+, this will
go away, so we just need to adjust our compat layer until then.

The compat layer was added to minimalize the code between the ADIN upstream
version and the ADI version.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 3023a0c3f1abdb69bebdd4b6bf6a02c315856668)