iio: jesd204: axi_adxvcri,xilinx_transceiver: pass QPLL selection directly from the state struct
This change moves the selection of the QPLL (0 or 1) to be an input of the
xilinx qpll functions. This API change is less error prone when configuring
the QPLL, as it is passed directly from the axi_adxcvr state struct (i.e.
st->sys_clk_sel).

This also means that the `conf` field from `struct xilinx_xcvr_qpll_config`
goes away, simplifying the use of this data-type. Using the `conf` field
was problematic as it would make the `struct xilinx_xcvr_qpll_config` have
a dual-role of both input & output when being used for QPLL functions.

The QPLL selection logic has been moved inside the xilxin_transceiver
library since that is something specific to it.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 7755910b41b0cae3401fb6f9f353c7b5974bc01a)