iio: ad9361: return the exact amount of bytes for gain-table
The reading of the gain-table would always return 4096 bytes. The trailing
data would be garbage.

This change limits the read to the exact amount of data that is read.

The issue was identified when investigating
   https://github.com/analogdevicesinc/libiio/issues/268

When doing:
```
iio_attr -a -d ad9361-phy gain_table_config > out
wc -m out
```

The value of `wc -m out` would be 4096. With the patch, the value is ~1688
bytes [i.e. the amount that is useful].

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 7c719442c2fb9875d73b4c9c7fc79a80195dfb8d)