iio: admc_speed: decouple from common AXI ADC definitions
This driver is a smaller copy of the AXI ADC driver. Since we want to
cleanup and evolve the main AXI ADC driver, it's a good idea to decouple
this old driver, which is used in some custom way in some old projects.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 83323200a1624f8c83b45379063338d641a46e60)